Reorganize helper functions
