+++ packages/web/src/app/map/utils.ts
@@ -1,39 +1,3 @@
-import { geojson as flatgeobuf } from "flatgeobuf";
-import { getFlatGeobufRectangle } from "../utils/get-flatgeobuf-rect";
-
-export async function getFgbData(map: any) {
-  let i = 0;
-  const geojson = { type: "FeatureCollection", features: [] as any[] };
-
-  const iter = flatgeobuf.deserialize(
-    "https://storage.googleapis.com/osm-tardis/2013-02-03T15%3A00.fgb",
-    getFlatGeobufRectangle(map),
-  ) as AsyncGenerator<any>;
-
-  const timestamps = new Set();
-
-  for await (const feature of iter) {
-    const { properties } = feature;
-
-    if (!properties || properties === null || properties === undefined) {
-      continue;
-    }
-
-    if (
-      properties.type !== "relation" &&
-      (properties.changeType === "added" ||
-        properties.changeType === "modifiedNew" ||
-        properties.changeType === "deletedNew")
-    ) {
-      geojson.features.push({ ...feature, id: i });
-      i += 1;
-      timestamps.add(properties.timestamp);
-    }
-  }
-
-  return { geojson, timestamps: Array.from(timestamps).sort() };
-}
-
 export function calculateStats(geojson: any) {
   const stats = {
     tags: {},
