+++ packages/web/src/app/utils/get-fgb-rect.ts
